contracts-bedrock: delete mainnet deployments
As part of https://github.com/ethereum-optimism/optimism/commit/6b1aa1e7de5d094954b6866f5605078b1ad81914
for some reason the `mainnet` directory was not deleted. This commit
deletes that directory as well. The source of truth for the deployments
is now in the `superchain-registry` repo.

When new contracts are deployed, their contract addresses can be found
there. Specific releases of smart contracts will contain the exact ABIs
for the deployed contracts at a release. These releases can be found
with git tags in the ethereum-optimism github repo.