added flag for hemitrap (#25)
* added flag for "hemitrap"

when testing private forks of heminetwork's production networks (i.e. testnet and mainnet), a few validations and features will need to be skipped (since the sequencer and other private keys/addresses change).  this pr adds a flag to op-node: "--hemitrap.enabled", when this flag is TRUE disable certain validations, features, and data that would hinder the progression of a forked network.

NOTE: this is NOT for production use

* remove accidentally added newline