do not override timeout in ForReceiptMaybe (#26)
optimism overrides context timeouts in (a lot of) test "wait" helper functions. one that we use use ForReceiptMaybe.  the timeout is hard-coded to 2 minutes.  but realistically when we "wait" sometimes what we're waiting for finishes in less than 2 minutes, sometimes longer.

remove this override.  use the context passed into the fuction for the timeout.